suggested artist can be fetched based on user prefered genre\

+++ kyoku-server/core/database/src/main/kotlin/com/poulastaa/core/database/entity/user/RelationEntityUserGenre.kt
@@ -5,7 +5,7 @@ import org.jetbrains.exposed.sql.Table
 
 object RelationEntityUserGenre: Table(name = "UserSavedGenre") {
     val userId = long("userId").references(EntityUser.id, onDelete = ReferenceOption.CASCADE)
-    val genreId = long("genreId")
+    val genreId = integer("genreId")
 
     override val primaryKey = PrimaryKey(userId, genreId)
 }
